Avoid compilation warnings on OS X 10.5 Leopard (https://trac.osgeo.org/gdal/ticket/2283)
git-svn-id: https://svn.osgeo.org/gdal/trunk@14046 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/ogr/ogrsf_frmts/gpx/ogrgpxdatasource.cpp
@@ -261,16 +261,16 @@ int OGRGPXDataSource::Open( const char * pszFilename, int bUpdateIn)
                 CPLError(CE_Failure, CPLE_AppDefined,
                         "XML parsing of GPX file failed : %s at line %d, column %d",
                         XML_ErrorString(XML_GetErrorCode(oParser)),
-                        XML_GetCurrentLineNumber(oParser),
-                        XML_GetCurrentColumnNumber(oParser));
+                        (int)XML_GetCurrentLineNumber(oParser),
+                        (int)XML_GetCurrentColumnNumber(oParser));
             }
             else
             {
                 CPLDebug("GPX",
                         "XML parsing of file failed : %s at line %d, column %d",
                         XML_ErrorString(XML_GetErrorCode(oParser)),
-                        XML_GetCurrentLineNumber(oParser),
-                        XML_GetCurrentColumnNumber(oParser));
+                        (int)XML_GetCurrentLineNumber(oParser),
+                        (int)XML_GetCurrentColumnNumber(oParser));
             }
             break;
         }
